+++ src/ska_mid_itf_engineering_tools/tmc_config/tmc_dish_ids.py
@@ -31,9 +31,19 @@ def instances(ids: str = "SKA000") -> list[str]:
     return instances
 
 
+def single_dish_id_uppercase(id: str = "SKA000") -> str:
+    """
+    Set DishID to uppercase for Tango device name.
+
+    :param id: string representing DishID, defaults to "SKA000"
+    :return: uppercase dishID
+    """
+    return id.upper()
+
+
 def single_dish_id_lowercase(id: str = "SKA000") -> str:
     """
-    Set DishID to lowercase for Tango device name.
+    Set DishID to lowercase for namespace device name.
 
     :param id: string representing DishID, defaults to "SKA000"
     :return: lowercase dishID
@@ -52,14 +62,14 @@ def dish_ids_array_from_str(ids: str = "SKA000") -> list[str]:
 
 
 def set_cluster_domain(
-    dish_id: str = "ska000",
+    dish_id: str = "SKA000",
     domain_postfix: str = "miditf.internal.skao.int",
     domain_prefix: str = "",
 ) -> str:
     """
     Make wild assumptions about the future of Kubernetes clusters in SKAO.
 
-    :param dish_id: Lowercase DishID, defaults to "ska000"
+    :param dish_id: Upper case DishID, defaults to "SKA000"
     :param domain_postfix: Cluster Domain where this dish is deployed, defaults to
         "miditf.internal.skao.int"
     :param domain_prefix: Prefix of Cluster Domain where this dish is deployed,
@@ -101,9 +111,10 @@ def dish_fqdns(
         dish_id: str = "SKA000",
         namespace_postfix: str = "",
     ) -> str:
-        id = single_dish_id_lowercase(id=dish_id)
+        id_upper = single_dish_id_uppercase(id=dish_id)
+        id_lower = single_dish_id_lowercase(id=dish_id)
         cluster_domain = set_cluster_domain(dish_id=id, domain_postfix=cluster_domain_postfix)
-        return f"tango://{hostname}.{namespace_prefix}{id}{namespace_postfix}.svc.{cluster_domain}:10000/{id}/elt/master"  # noqa E501
+        return f"tango://{hostname}.{namespace_prefix}{id_lower}{namespace_postfix}.svc.{cluster_domain}:10000/mid-dish/dish-manager/{id_upper}"  # noqa E501
 
     fqdns = [
         single_dish_fqdn(